fixed an ordering problem

+++ WhirlyGlobeSrc/WhirlyGlobeLib/src/SceneRendererES2.cpp
@@ -343,7 +343,7 @@ void SceneRendererES2::render()
 //            [activeModel updateForFrame:&frameInfo];
         
         if (perfInterval > 0)
-            perfTimer.addCount("Scene changes", scene->changeRequests.size());
+            perfTimer.addCount("Scene changes", (int)scene->changeRequests.size());
         
 		// Merge any outstanding changes into the scenegraph
 		// Or skip it if we don't acquire the lock
